requirements.txt
tensorflow>=2.4.0
dm-sonnet>=2.0.0
absl-py>=1.0.0
numpy>=1.18.0
matplotlib>=3.1.0
streamlit
numpy
pandas
pyvista
vtk
matplotlib
scipy
plotly
trimesh
tqdm
scikit-learn
absl-py
psutil
nvidia-ml-py3